requirements.txt
flask
ultralytics
pillow
numpy
